+++ ios/Sources/NativePermissionsPlugin/Models/PermissionStatus.swift
@@ -4,6 +4,7 @@
 
 internal enum PermissionStatus: String, CaseIterable {
     case granted = "granted"
+    case restricted = "restricted"
     case denied = "denied"
     case permanentlyDenied = "permanently_denied"
 }