+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 1.0.0
+
+- Add an option when opening app settings to wait until the user returns to the app
+
+#### iOS
+- Return restricted status when appropriated
+- Add: `showRationale` method to optionally display relevant information to users before directing them to the app settings
+- Fix: ensure denied is return across all permissions when returning from denied prompt
+
 ## 1.0.0-beta.1
 
 - Initial release